+++ .golangci.yml
@@ -1,0 +1,8 @@
+run:
+  timeout: 5m
+issues:
+  fix: true
+linters:
+  enable-all: true
+  disable:
+  - exhaustivestruct
